requirements.txt
aiofiles==0.8.0
python-dotenv==0.20.0
Telethon==1.24.0
